Redirect non existing URLs to the root
The UI generates URLs like `/clusters` which don't correspond to any
real resource in the web server. If the user then clicks the reload
button the browser will send this URL to the server, which will return a
404 error response. To avoid that this patch changes the web server
configuration so that the web server will instead redirect to the root
of the application.

Signed-off-by: Elad Alfassa <ealfassa@redhat.com>