Add build identifier
This patch adds two new `Makefile` variables to identify the build:

- `build_id` - Uniquie identifier of the build.
- `build_ts` - Timestamp of the build.

The values of these variables are currently only used in an init
container that simply prints their values. This is intended to force an
automatic rollout of the new version of the application when it is
deployed from Jenkins.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>