Get versions from API if available
Currently the _OpenShift_, operating system and container runtime
version fields in the cluster details view are always `N/A`. This patch
changes the application so that it will get it the values from the API.
If the values aren't returned by the API, or if they are empty, the
application will still display `N/A`.

Related: https://jira.coreos.com/browse/SDA-109
Signed-off-by: Juan Hernandez <jhernand@redhat.com>