Get creation timestamp from API if available
Currently the _CREATED AT_ field in the cluster details view is always
`N/A`. This patch changes the application so that it will get it from
the `creation_timestamp` returned by the API. If the attribute isn't
returned by the API, or if it is empty, the application will still still
`N/A`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>