Update to `yarn` 1.10.1
This version of `yarn` adds to the entries in the `yarn.lock` file a new
`integrity` attribute that contains a cryptographic digest of packages,
to improve security. Without this patch each time that I buid the
application git wants to add the `yarn.lock` changes to the commit,
which creates a lot of noise. It also possible to disable this adding the
`--skip-integrity-check` to `yarn` command, but it is better to just
update the `yarn.lock` file.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>