ClusterList, ClusterDetails: Use cloud provider name from the API
In ClusterList, we need the "short" version of the provider name,
so we just use the id field and transform it to uppercase.

In ClusterDetails we need the "long" version of the name
(display_name from the API), but doing an extra request just for
this tiny detail will not be the best user experience.

This commit uses the short version of the name in both screens,
in hope that the API can be changed to provide display_name
when asking for cluster details.

If the API won't change, I'll implement the extra request in a
followup PR.