Many lint fixes
This commit makes the PropTypes lint rules a bit less strict,
as it's too much work to specify the shape of every API object
we use (and it will certainly get out of sync with the backend).

In addition to that, the commit fixes lint errors all over the
tree, down to 17 lint issues.

The final goal is to get to 0 lint issues, and then enable
eslint verification in Jenkins, to make sure we don't
miss lint errors when reviewing merge requests