Configure jest to work without react-scripts
Some work was needed here because we need to use babel to
transform our source code, but we usually configure babel
from inside webpack.
This was solved by having babel.config.js, set up with the
transformation presets only for the test env, and not
for building the app.

It needed to be babel.config.js and not .babelrc for jest
to be able to use it properly.

Some extra tweaking was needed in package.json to make sure
lodash-es is preprocessed by babel-jest (which usually ignores
node_modules).