Add portal launcher
Currently the configuration of the Apache server is generated inside a
configuration map, which is then mounted by the server container. But
this doesn't work well with `saasherder` because it doesn't support
configuration maps. To avoid using that configuration map this patch
adds to the project a launcher, written in Go, that generates the
configuration and starts the Apache server.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>