Mark project `.eslintrc` as root
The `eslint` tool by default uses all the `.eslintrc` files that it
finds in the project directory, in its parent, etc. This complicates
preparing the project for building Go code in Jenkins, because that
needs to create a `.gopath` subdirectory that contains a copy of the
project that satisfies Go tools. That copy would contain also the
`.eslintrc` file, resulting in two identical `.eslintrc` files that will
be merged together. Unfortunately that results in an error. To avoid
that this patch adds `"root": true` to the `.eslintrc` file, so that the
the search will stop at the first file.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>