Use `IMAGE_TAG` like `saasherder`
The `sassherder` tool has a built-in variable named `IMAGE_TAG` that
contains the last part of the image name, usually the short git hash.
But in this project it contains the complete image name and tag. To
resolve that conflict this patch adjust the project so that it uses
`IMAGE_TAG` in the same way that `sassherder`.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>