ClusterList: only fetch the cluster list once on load
Use componentDidUpdate instead of the deprecated componentWillUpdate,
and fix a problem where the cluster list was fetched twice on load.

The issue happened because the cluster list starts with valid=false,
and once the request is sent, the state updates to say that it's
pending. Then, valid is still false, so fetch would've been called
again.

Making sure to not refresh on invalidation when a request is
pending fixes this issue.

https://jira.coreos.com/browse/SDA-277