package.json: Add path-to-regexp to dependencies
I've noticed an error which prevented the UI from rendering because
sometimes yarn would pull a much older version of path-to-regexp,
(a dependency of some of our dependencies) which broke react-router.

Searching for the error lead me to this github issue,
https://github.com/ReactTraining/react-router/issues/6203
which mentioned path-to-regexp should be added to package.json to
work around the issue.