Remove old navigation
In the previous sprint, we hid the navigation bar using a CSS hack,
because the fate of the navigation was not decided yet.

Since then, the fate of navigation has been decided: if the user
goes through the try.openshift.com flow, they should get no navigation,
and if the user comes through cloud.openshift.com, they should
get navigation.

However, there's no point to show the user any navigation until
we actually have more screens to navigate to.

This commit removes the old navigation code, so that there's no
longer a big gap on the left side on the screen, and the code
is much cleaner.

We'll add navigation back when we'll add more screens.