instructions: Drop "extract" and add Terraform download
The installer requires the 'terraform' command in the PATH (a few
other places work too, but the PATH is the recommended approach).
Previous versions of the Tectonic Installer bundled a 'terraform'
binary in a release tarball, but the OpenShift Installer releases are
just openshift-install binaries [1]; users have to bring their own
Terraform [2].

Also update the documentation link to point at the quick-start.  The
docs/ dir has additional details beyond the quick-start, but the
quick-start is the intended landing page that links out to pages with
more detail.

Putting <code> on its own line avoids [3]:

  /var/lib/jenkins/workspace/service-uhc-portal-gl-pr-check/.gopath/src/gitlab.cee.redhat.com/service/uhc-portal/src/components/clusters/install/components/Instructions.jsx
    121:32  error  `.              Run the installer ...` must be placed on a new line  react/jsx-one-expression-per-line

Docs for the check are in [4].  Putting the trailing period on the
next line feels strange, but the JSX docs have [5]:

  JSX removes whitespace at the beginning and ending of a line. It
  also removes blank lines. New lines adjacent to tags are removed;
  new lines that occur in the middle of string literals are condensed
  into a single space.

We avoid the "condensed into a single space" case here because the
newline is adjacent to a tag and at the beginning of the string
literal.

Also fix some tab/space inconsistency in the region of the new
Terraform URL plumbing.

[1]: https://github.com/openshift/installer/releases
[2]: https://github.com/openshift/installer/blame/master/README.md#L25-L29
[3]: https://ci.int.devshift.net/job/service-uhc-portal-gl-pr-check/31/console
[4]: https://github.com/yannickcr/eslint-plugin-react/blob/v7.11.1/docs/rules/jsx-one-expression-per-line.md
[5]: https://reactjs.org/docs/jsx-in-depth.html#string-literals-1