instructions: Format the token as a single line
The token has been pretty-printed with two-space indents since it
landed in 854233ff (Provide MVP tollbooth functionality, 2018-10-15).
But the installer's prompt system expects input to be a single line
[1].  Adjusting the "Copy Pull Secret" payload to use a single line
addresses most of that.  We could leave the "Download Pull Secret"
content pretty-printed, but there's a chance that users could download
that and then copy/paste from the downloaded file into the installer
prompt, so I've just dropped pretty-printing for both instances.

[1]: https://github.com/openshift/installer/pull/629