getErrorMessage: gracefully handle JS Error objects
QE noticed that the cluster list gets stuck forever if
a user logins with non-redhat emails. This looked weird
to me, because we have error handling code.

Testing this, I've modified the backend to always return
an error - and couldn't reproduce. So I deployed the API
gateway locally, and changed the ACLs to make it block
everything, and that reproduced the error.

Turns out that Apache does not return the access-control-allow-origin
if authentication fails, so while the OPTIONS call succeeds and
has the right headers, the error page does not. This caused
Firefox to treat this request as a CORS failure, and the payload
to getErrorMessage was a JS error object saying that this is a
"Network Error", and the getErrorMessage would've raised
an exception, that wouldn't be caught, so the spinner
kept spinning forever.

While still not addressing QE's original concern (show something
nicer to people with non-Red Hat emails), showing some kind
of error message is still better than a never ending loading screen.

Related: https://jira.coreos.com/browse/SDA-354