EditDisplayNameDialog: Don't invalidate cluster list on cancel
EditDisplayNameDialog expects a closeFunc, which gets a boolean
argument which should specify if there was an update or not.

Calling closeFunc directly from the close button click event
caused that argument to contain the click event, which
evaluates as true. This would cause the cluster list to
always get invalidated when the dialog is closed.

Using an anonymous function to call closeFunc ensures this
argument will be false, and the cluster list won't
get invalidated.

https://jira.coreos.com/browse/SDA-357