Fetch cloud provider and region information
This MR fetches cloud provider and region information from the
backend, introducing a new reducer to put this info in the redux
state, and a component to display it in the cluster list,
and a component to display them in the cluster creation form.

Before region and cloud provider info is loaded, the IDs are displayed,
so the user doesn't have to wait for information to be loaded before
it's displayed.