createClusterForm: fix submit button
The button stopped working when the structure was changed and
the form was put inside the modal body. This caused the submit
button to not be a child of a form, and so submission did nothing.

redux-form documentation says handleSubmit can be called either
from onSubmit on the form, or from onClick on the button,
so I moved it to onClick.