CreateClusterForm: Add a loading spinner, disable fields when loading
Requests to create a cluster sometimes take a second or two to complete.
While this happens, we should let the user know we're processing their
request, otherwise it might look like "nothing happened" until
the user is redirected to the cluster details.

The pattern I used is based on the Patternfly form examples.