ClusterList: Change create cluster button to dropdown
Allowing to pick OCP or OSD.

Ideally we should change this to show the /clusters/install screen
in the modal, but I don't want to add more modals or change them
before we figure out how to do this cleanly (we've got way too many
modals and the code has gotten messy and duplicated) so this
is probably good enough for now.

I don't like the fact that the link to /clusters/install causes a
reload,
but I don't know how to fix it - using <Link> from react-router
causes only part of the menu item to be clickable.

https://jira.coreos.com/browse/SDA-342