Reduce getCloudProviders latency and avoid multiple requests
calling getCloudproviders() before the first render reduces
the time it takes for the user to get the cloud provider list,
since it doesn't wait for the render to make the request.

This reduces flicker on the cluster list, as the region display
name will be loaded sooner.

Additionally, don't call getCloudProviders() unconditionally
on ClusterList's componentDidMount() - that call was there
to avoid multiple calls being fired from all the ClusterLocationLabel
objects being rendered at the same time, but having it there
with no condition made it fire even when we already had the
data, for example if the user loaded the app, moved to the
cluster details screen, and then went back to the cluster list.

https://jira.coreos.com/browse/SDA-438