clustersActions: Add tests
These tests check that the action dispatches correctly and
calls to the right clusterService function to make the request.

To implement the check that calls clusterService, I've mocked
it in such a way that all the mock promises return functions
that resolve to empty objects. Since the actions here don't
do any processing of the results, this mock is good enough
for now.