Automatically deploy route
Currently the route for the portal isn't automatically created by `make
deploy`. The reason for that was that the route contained secret
information, in particular the TLS private key. But with the change to
use _saasherder_ to deploy to the production environments it is no
longer necessary for the _Makefile_ in this project to take care of
that. This patch takes advantage of that, making changing the _deploy_
target of the _Makefile_ so that it will also deploy the route. This
reduces the steps to perform manually in the development environments.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>