main.jsx: Expire the session after 15 minutes of idle time
When the user is idle for 15 minutes, log them out.
Idle time is calculated by checking the last timestamp the user clicked
a button, used the mouse scrollwheel, pressed a key on the keyboard,
or touched the touchscreen.

If the user is active, we keep refreshing the keycloak token as usual,
but we actively log them out if they're idle.

The timestamp is kept in localStorage, to support a case in which
the portal might be opened in separate tabs - so that activity in
one tab will reset the idle counter for all tabs, preventing unexpected
logouts.

https://jira.coreos.com/browse/SDA-309