Don't show the alpha notice in the try.openshift.com flow
Currently the alpha notice is shown in all the pages of the application,
including the try.openshift.com. This patch changes it so that it will
only be visible in the pages that are intended for internal use, and not
visible in the page that contains the instructions to install a cluster
manually.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>