Add backend proxy
This patch adds a backend proxy server intended for development. The
`make binaries` command generates a new `backend` binary that can then
be used like this:

```
$ ./backend
```

That listens in `https://localhost:8002` and acts a proxy for the
_OpenID_ and API services of the staging environment.

The backend supports additional request forwarding configurations, using
a custom configuration file and the `proxies` element. For example, to
forward requests to a local clusters service, while sending the rest to
the staging environment, the user can create a `my.yml` configuration
file like this:

```
proxies:
- prefix: /api/clusters_mgmt/
  target: https://api.127.0.0.1.nip.io
```

And then run the server with this command:

```
$ backend --config=my.yml
```

The patch also changes the default configuration of the we application
so that it uses this backend proxy.

Related: https://jira.coreos.com/browse/SDA-558
Signed-off-by: Juan Hernandez <jhernand@redhat.com>