Add mock data + mock backend
This simple mock backend only supports read-only operations,
and doesn't support searching/sorting/paging,
but it is enough to get the UI running even when the real backend
is not working.

This will simplify UI development and prevent cases in which we're
blocked on fixing bugs / developing new features when the staging
API server is broken. It's easily extendable to support new API
entities, and allows testing the UI behaviour on different types
of clusters without actually creating any real cluster on AWS.

The `mockdata` directory structure follows the API path structure.
An object which is readable via the API should have a json file
with the json we expect the API to return.