ClusterDetails: Add breadcrumbs
The breadcrumbs replace the "Back" button. Note that these are
temporary, static breadcrumbs, and we will probably need to replace them
with something more complex when we have more screens.

This patch also creates a new `<ClusterBadge>` component to reduce code
duplication with the badge definition, adds the badge to the
ClusterDetails screen, and moves the refresh button to the right side of
the screen in ClusterDetails, to match its location in the ClusterList.

This fixes https://gitlab.cee.redhat.com/service/uhc-portal/issues/5