ClusterDetails: Fix calculation of "available" memory/storage
The previous calculation just assumed humanization would output GiB.
This assumption is obviously wrong, since humanization can output
smaller or bigger units, or output different units for used/total.

Because of that, the chart component needed to be changed to allow
either values with pre-defined units (such as CPU) or byte values
which would be humanized inside the donut component instead of outside
of it.

This allows the "available" calculation to happen in bytes, and then
humanized for display on the donut.

https://jira.coreos.com/browse/SDA-565