ClusterDetails: mark the view "pending" if we have data for the wrong cluster
The clusterDetails reducer keeps the previous data it had even when
requesting a new cluster. This makes sense when you want smooth refresh
without replacing the view with a loading screen.

As a side effect, this meant that if you opened the details for one
cluster, got back to the list view, and opened another, for a split
second the details for the previous cluster would be rendered.

As well as being weird for people with slow connections, it also
uncovered a bug with the refresh button that made it behave weirdly -
if it was unmounted and then remounted too quickly, accessing the state
from componentWillUnmount proved unreliable, causing a timer leak.