ClusterDetails: add a spinner when the view is refreshing
This adds a small spinner next to the cluster name when the cluster
details page is refreshing (either automatically or after you clicked
the button).

Note that the spinner appears only for a very short amount of time,
unless your connection (or our backend) are very slow.

This should address https://jira.coreos.com/browse/SDA-714