Show organization name in the masthead
This MR adds two calls to the accounts service to find the name
of the user's organization, and then show it after the user name in the
masthead.

It fixes https://jira.coreos.com/browse/SDA-736
also fixes #7
and also helps a bit with https://jira.coreos.com/browse/SDA-700 but doesn't
fix it.