EditDisplayNameDialog: Stop using redux form
Using redux form for a single-field form is a bit of overhead
that we don't really need, and has some weird bugs caused by
initialValues weirdness and the fact that the modal is always mounted.

Instead of spending hours trying to figure out how to use initialValues
correctly in a case of a form that effectively gets recycled, porting
it out of redux-form and into a simpler form is easier.

This should fix https://jira.coreos.com/browse/SDA-765

also, add tests.