Add `token` page
This patch adds a new `/tokens` page to the application that displays
the access and refresh tokens in a way easy to consume for users of
the API from the command line, or from non-browser applications.

For users of the API it is convenient to have refresh tokens that never
expires, and that requires adding the `offline_access` scope to the
login request. That will be done automatically when the user types
`.../token` in the browser address bar.

Related: https://jira.coreos.com/browse/SDA-480
Signed-off-by: Juan Hernandez <jhernand@redhat.com>