Use token in the backend proxy server
This patch changes the backend proxy server so that it uses an offline
access token for authentication instead of an user name and password.

An important side effect of this is that the proxy server can now only
authenticate one user: the owner of the offline access token. That log
in will happen automatically without asking for user name or password.

By default the proxy server expects the offline access token in the
`UHC_TOKEN` environment variable, or in the `token` parameter of the
configuration file. It will not start if it doesn't find it.

Related: https://jira.coreos.com/browse/SDA-480
Signed-off-by: Juan Hernandez <jhernand@redhat.com>