Keycloak: refresh token more often
Beni and Oved noticed an issue where they'd get a "Network Error"
if they kept a tab open on Cluster Details for too long.

Beni and I figured out that the minValidity parameter for Keycloak
was too low - it meant that the token will only be refreshed if
there were only 10 seconds (or less) of validity left. Combined with
the 10 seconds refresh interval it meant that there could be,
in some very unlucky edge cases (that get more and more likely
the longer the tab is open), a window of *one second* for the
token to be invalid before it's refreshed again.

One second is clearly not enough.

This commit changes the minValidity parameter to 60 seconds, meaning
that the token will be refreshed if it's going to expire in 60 seconds
or less. It also increases the token refresh frequency to make
cases like this even more unlikely.