Remove short lived access token
Currently the tokens page shows the short lived access token and the
offline access token. This causes confusion, and doesn't make it
completely clear that the token that should be used with the API is the
offline access token. To reduce that confusion this patch removes
completely the short lived access token from the page, and changes the
wording so that it always refers to the offline access token with the
_offline_ word.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>