Tokens page: Support embedded mode (cloud.redhat.com)
When we're embedded on the Insights platform at cloud.redhat.com,
we can only have one HTML page, so the trick with a separate token.html
will no longer work.

Furthermore, we don't control keycloak in that mode, so we needed
another solution.

Ian was kind enough to add an API for us to request an offline token
when we're in embedded mode, keeping most of the complexity inside
the insights chrome.

This MR uses this API to allow the tokens page to work in embedded mode.
To get to the token page in embedded mode, you need to navigate to
/token instead of /token.html

This MR depends on https://github.com/RedHatInsights/insights-chrome/pull/256
to function, but it should be mergeable before that lands, because
it's broken on master anyway and it can't get "more broken".

Related: https://jira.coreos.com/browse/SDA-771