RefreshButton: turn off autoRefresh when hidden or offline
Based on feedback in uhc-feedback, people got confused when they
switched networks or undocked their laptop and came back to the UI
showing an error. Checking the "online" state should avoid refreshes
in this specific case.

In addition, refreshing when the tab is hidden is needless load
on the backend, so that should be disabled as well.

https://jira.coreos.com/browse/SDA-848