Don't use `8.8.8.8` in instructions page
Currently the instructions page suggests the user to use `8.8.8.8` as
the DNS server for testing that DNS is working. This isn't required in
general, and it doesn't work from within the Red Hat VPN in particular
because access to DNS servers other than the internal ones is blocked.
This patch removes the `@8.8.8.8` part from the `dig` command.

Related: https://jira.coreos.com/browse/SDA-811
Signed-off-by: Juan Hernandez <jhernand@redhat.com>