main: properly find and store keycloak token when chromed
I apparently didn't actually test this works because I couldn't
get the local insights chroming setup to work with the backend proxy,
so I only tested with mock mode, which doesn't require auth.

This MR fixes this, and now properly stores the keycloak token
in the session storage so we can add it to API requests.

Note that once we're 100% chromed and we'll
remove the option to run non-chromed, we should find a better way
to pass the token to requests (because it doesn't make much sense
to store it in the session storage if we're not the ones managing
authentication).