Different builds for different branches
Currently we only push the build to the `*-beta` branches of the
Insigths platform. The reason is that we only do a build, and it
generates an `index.html` that has `beta/` in the links to the assets.
This patch changes the build script so that it does two builds, one with
the `beta/` prefix and another without. The builds with `beta/` will be
pushed to the beta branches, and the buidlds without will be pushed to
the stable branches.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>