services: Get a token from Insights on every request
This MR changes the way we do all API calls. Now, instead of
calling axios directly, it's called from a wrapper, so we'll be
able to always attach the right token to the request.

This is needed because when running chromed on the Insights platform,
we don't control token renewal, we have to ask for a fresh token
from Insights, which is asynchronous.

So for the non-chromed flow, we keep doing what we always did -
get the token from the session storage, and trust our refresh timer
to put it there. But when we're chromed, we call the Insights
getToken method on every request.

Without this MR, we'll only have a valid token for the first 5
minutes, and then it would expire and things will be broken.