Remove router shard schemes
Internal router shard schemes are not supported on v4 clusters, and even
when supported, there is no way to update them in OpenShift[1]. For now
I am removing all instances of the scheme form field and hard-coding it
as internet-facing.

Also added an editedRouterShard reducer to track when they get modified
in the EditClusterDialog.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1686206#c3