HeaderToolBar: smarter user display name fallback logic
This is required for sso.redhat.com which doesn't always have
preferred_username.

We also go the extra step in the fallback logic and fall back to
the user's username or email so that there will always be *something*
useful in this field, even if they somehow don't have a name set.