Fetch organization for the insights flow as well
The cluster list code assumes that the organization will be fetched,
but it wasn't fetched in the insights flow because we didn't need
to show it on the masthead (since we don't control the masthead).

This MR fixes it to fetch the organization on the insights flow
as well. Without this MR the insights flow will be stuck on loading
forever.