Cluster List: make errors less fatal
If we have fetched clusters successfully at least once,
make any further error non-fatal. Instead of replacing the entire
view with an error message, show a small triangle in place of the
spinner, with a tooltip that shows the error message.

Refreshing (and auto refreshing) is still possible when such
errors occur.

To test you can load the cluster list, and then disconnect
your network (or close your backend proxy), and try clicking the
refresh button.