Cluster List: fix mock mode, and fix quota pending logic
If loading quota or organization failed, don't wait forever,
just assume no quota.

Also, Cluster List requires the getOrganization request to
be dispatched, so make sure that it is dispatched
even if `main.jsx` didn't dispatch it for whatever reason
(for example, if we're in mock mode)

This MR restores mock mode functionality. Future MRs need to add
mock data for quotas so we can test the create cluster modal
again in mock mode.