EditDisplayName: behave as expected when hitting Enter in the textbox
When focused on the display name text box and hitting Enter, the form
tried to submit, presenting the user with a weird and unexpected error
message.

This doesn't happen for DeleteClusterDialog because it does not
have a `<Form>`, and it doesn't happen on `EditClusterDialog`
because it's managed by Redux.

Adding an onSubmit handler with `event.preventDefault()` prevents
the default browser form submission from happening and as a nice
bonus also allows users to press Enter to edit the display name,
instead of having to click the Edit button.