Script to push manually to Insights
This patch adds a new `push_to_insights.sh` script that builds the
application and pushes it to the production branches of the Insights
deployment repositories.

To use the script you will need the private key that grants permission
to push to that repository. That key should be encoded using Base64, in
one line, and the value should be in the `PUSH_KEY` environment variable:

```
PUSH_KEY=$(base64 --wrap=0 the_key_file) ./push_to_insights.sh
```

Signed-off-by: Juan Hernandez <jhernand@redhat.com>