ClusterList: modify "loading message" condition to account for validity
When we delete a cluster, we invalidate the clusters list to make
it refresh. The condition used to check if the list is pending
and there's no filter to show the loading message, and that worked
okay for most cases,

but when we delete a cluster and then go back for the cluster list
with a filter set, it would've first rendered "no results"
(because the cluster list in memory was empty), then it'd call refresh
and the loading indicator would appear.

Checking validity in this condition ensures we show the loading
message in this case, and not "no results".

https://jira.coreos.com/browse/SDA-893