clusterStates: fix "installation is taking too long" check
The "installation taking too long" was meant to display the
"installation taking too long" message if the installation takes
*more than* two hours, but because getTimeDelta rounds up, this meant
that one hour and one minute would be caught by the >= 2 condition,
and the "installation is taking too long" message would be displayed.

Changing the condition to > 2 ensures that the message will not
appear too early.

This is not a critical bug, because one hour and one minute is
already quite long for installation, but it's probably good to
fix it anyway.

https://jira.coreos.com/browse/SDA-983