Users screen: Add informative help text
I have a feeling that it's not clear enough for people unfamiliar with
the system what the Users screen does. I added an informative text to
try to make this more clear for new users.

Also, I made the spinner spin while waiting for the user list,
to address feedback from Beni on the original Users screen MR.