fix responsiveness issues for install page cards
switch the /install and /aws pages from css grid to flex
to correct the mobile responsiveness bug where the page
was cut off. Now everything is 2 cards across on normal /
large screens, and 1 single card on small screens with
the rest of the cards flowing below.

/install is now two stacked flex containers to preserve
the 2x2 layout when there's enough width for 3+ cards to
flex in.