ClusterDetails: Reset identity provider state on unmount
When we open the details for one cluster, and then move to another,
identity providers will only be fetched once - for the first cluster.

That's because the cluster details didMount handler thinks that the
request is already fulfilled, so it doesn't bother making the request
again.

Resetting the state when the component is unmounted ensures that
identity providers will be fetched again when moving between clusters.

This should fix https://jira.coreos.com/browse/SDA-990