ClusterDetails: Use the same error triangle pattern like clusterList
This allows errors caused when refreshing to not be "fatal", by
displaying a warning triangle on the top row instead of replacing
the entire screen with an error message.

This patch is a first step of a two step fix, the next step would be
to check if the error was 404, and if it was, redirect the user back
to the Cluster List, and show an error on top of it saying the cluster
might've been deleted.

https://jira.coreos.com/browse/SDA-809