Redirect from Cluster Details to Cluster List upon a 404 error
This MR introduces the concept of "global errors". Global errors
can be set from any components, and will be displayed by any component
that opts to display them. They're stored in the redux state, so they
don't get reset when a component is unmounted.

Using this concept, we can redirect from the Cluster Details screen
back to the cluster list if the user tried to navigate to a cluster
that doesn't exist, and display the relevant error on top of the cluster
list.

This might happen if the user clicked a cluster that was already
deleted, before their cluster list was refreshed.

To try it out, navigate to the details of a cluster that
does not exist, for example /details/foo

This fixes https://jira.coreos.com/browse/SDA-809