Properly handle empty organizations
According to the SD-B team, when an account is first created,
it might have no organization. This caused organization.id to be
undefined, which caused an unnecessary request to be sent to fetch
an organization with the ID "undefined", which is not a huge issue,
but it's not very nice either.

This commit makes sure we don't send a request for organization
when the organization ID is undefined.

https://jira.coreos.com/browse/SDA-1121