Add Operation ID to all errors
The server now returns operation ID for errors, so we should show it to
the user, in hopes that they'll include this identifier in bug reports
and it'll make deubugging the backend easier.

To do this, I had to change all reducers, so I created a helper function
to parse the error object returned from the server and return a
structured error state fit for redux, saving a lot of reducer
boilerplate.

Also, since we're already modifying all alerts, I moved them all to a
component that picks the relevant fields out of the response, so next
time we need to change something with alerts, we'll only need to do it
in one file, not ALL files.

And since I already created this component, it's a good opportunity to
move (most of) the alerts to Patternfly 4

Fixes https://jira.coreos.com/browse/SDA-938
Related https://jira.coreos.com/browse/SDA-1153