Cluster Creation Dialog: Add node type selection
This MR adds a new MachineTypeSelector component to select the machine
type when creating a cluster. It shows a list of radio buttons for all
the types.

The radio boxes are implemented in PF4, so we won't need to do the work
twice of creating them now and then converting them from PF3 to PF4.

To achieve this it includes some CSS hacks to make the radio buttons fit
the existing PF3-based form, which will need to be removed later.

For the first step, this still doesn't take quota into account, so it'll
show you all the options even if you can't use them. Since this feature
is urgent, we should first finish this step, and then add the quota
support later.

https://jira.coreos.com/browse/SDA-1130