CreateCluster: fetch quota when needed
Right now, the only screen that handles fetching quota is the cluster
list. This means that if you navigate to /create directly, it won't
fetch quota, and show you just one card (for OCP).

This is really not great, because the experience should be consistent
no matter which page you navigated to first.

It's also not great because the logic to fetch the organization and the
quota is handled by the component itself instead of being abstracted
away, and this means that doing this from two components requires
duplicating that logic.

This is an urgent quick fix (since this problem affects prod too)
but later we should refactor this so that the redux actions handle
fetching both the organization and the quota without having to have
any special logic for both in the component lifecycle methods.