Machine type selection: disable types according to quota
This MR adds the functionality for the cluster creation form
to check the user's available quota, and disallow selecting
node types that the user has no quota for.

It also ensures that the machine types are sorted by size, so
the list is a bit less confusing.

Fixes: https://jira.coreos.com/browse/SDA-1234

Related: https://jira.coreos.com/browse/SDA-1130