Allow environment overrides
For the purpose of testing the front-end with different backends
on-the-fly, we want to be able to send a query param of the type
`?env=$ENV`, where `$ENV` is one of the environments configured in the
`src/config/config.json` file. This will override any global
configuration with the one from the specific environment.

This configuarion is not available in the production deployment.