Machine types: Implement the new design
To implement this new design, I've created a new component
FlatRadioButton, that is based on a PF4 button, but styled flatly,
and has a "selected" state. It needs to be a separate component because
we'll soon need it for the "red hat infrastracture" vs "byoc"
selection.

This also implements setting a default MachineType automatically -
the default is set to the first MachineType that has quota.

Fixes: https://jira.coreos.com/browse/SDA-1235