Remove all router shard functionality
Customers should not be able to set router shards when creating an
OpenShift Dedicated cluster. All clusters currently are deployed with
'apps' as their default/public router shard (and corresponding
IngressController). In the future (4.2+), we will allow customers to add
an additional router shard if public and private routes are needed on
the cluster. For now, we need to remove these provisioning options.

Also rename "Edit Cluster" to "Scale Cluster".